Preserve uninitialized Original AI old-state words
Linux-v48 stores full four-byte enum records. The apparent packed-enum values came exclusively from RHArtificialIntelligence::mOldState, which its constructor never initializes and Original never reads; StartThink is its only assignment. Preserve that exact signed storage word instead of fabricating an AiState, and overwrite it from current_state at the Original pre-filter Think boundary.

Remove the incorrect signed-low-word masking from behaviorally consumed AI enums so malformed full-width values fail strict validation. Corpus examples such as 0x00960003 and 0x5a3b010e remain inspectable and state-hashed exactly. Add focused coverage for indeterminate old-state storage.

+++ crates/robin_engine/src/ai_enemy/mod.rs
@@ -3125,7 +3125,7 @@ impl EnemyAi {
         self.base.couldnt_reachpoint = false;
         self.base.already_on_point = false;
         self.base.already_turned = false;
-        self.base.old_state = self.base.current_state;
+        self.base.old_state = self.base.current_state as i32;
         self.base.think_recursion_depth += 1;
 
         // Track stimulus actor
